feat: Add bulk swatch handling and logging functionality in LabDip ZipperSwatch
