+++ src/pages/LabDip/ZipperSwatch/index.jsx
@@ -1,22 +1,17 @@
 import { useCallback, useMemo, useState } from 'react';
 import { useAuth } from '@/context/auth';
-import { useDyeingSwatch } from '@/state/Dyeing';
+import { useDyeingSwatch, useDyeingSwatchBulk } from '@/state/Dyeing';
 import { useOtherRecipe } from '@/state/Other';
-import { format } from 'date-fns';
 import { useAccess } from '@/hooks';
 
 import ReactTable from '@/components/Table';
-import SwitchToggle from '@/ui/Others/SwitchToggle';
-import {
-	DateTime,
-	LinkWithCopy,
-	ReactSelect,
-	StatusButton,
-	StatusSelect,
-} from '@/ui';
+import { StatusSelect } from '@/ui';
 
 import GetDateTime from '@/util/GetDateTime';
 import PageInfo from '@/util/PageInfo';
+
+// Import the column configurations
+import { createColumns, createColumnsLog } from './columns';
 
 const options = [
 	{ value: 'all', label: 'All' },
@@ -31,7 +26,6 @@ const options2 = [
 
 export default function Index() {
 	const { user } = useAuth();
-
 	const [status, setStatus] = useState('pending');
 	const [status2, setStatus2] = useState('incomplete_order');
 
@@ -40,14 +34,24 @@ export default function Index() {
 			? `order_type=${status2}`
 			: `type=${status}&order_type=${status2}`
 	);
+
+	const { data: bulkSwatch, isLoading: isLoadingBulkSwatch } =
+		useDyeingSwatchBulk(
+			status2 === 'complete_order'
+				? `order_type=${status2}`
+				: `type=${status}&order_type=${status2}`
+		);
+
 	const info = new PageInfo(
 		'LabDip/ZipperSwatch',
 		'order/swatch',
 		'lab_dip__zipper_swatch'
 	);
 	const haveAccess = useAccess('lab_dip__zipper_swatch');
+	const { data: recipe } = useOtherRecipe(`approved=true`);
 
-	const handleSwatchStatus = useCallback(
+	// Handler functions
+	const handleSwatchStatusLog = useCallback(
 		async (e, idx) => {
 			await updateData.mutateAsync({
 				url: `/zipper/sfg-swatch/${data[idx]?.uuid}`,
@@ -60,320 +64,135 @@ export default function Index() {
 		},
 		[updateData, data]
 	);
-	const { data: recipe } = useOtherRecipe(`approved=true`);
 
-	const handleSwatchApprovalDate = async (idx) => {
-		await updateData.mutateAsync({
-			url: `/zipper/swatch-approval-received/${data[idx]?.uuid}`,
-			updatedData: {
-				swatch_approval_received: data[idx]?.swatch_approval_received
-					? false
-					: true,
-				swatch_approval_received_date: data[idx]
-					?.swatch_approval_received
-					? null
-					: GetDateTime(),
-				swatch_approval_received_by: user?.uuid,
-			},
-			isOnCloseNeeded: false,
-		});
-	};
-
-	const columns = useMemo(
-		() => [
-			{
-				accessorKey: 'order_number',
-				header: 'O/N',
-				cell: (info) => {
-					const { order_number } = info.row.original;
-					return (
-						<LinkWithCopy
-							title={info.getValue()}
-							id={order_number}
-							uri='/order/details'
-						/>
-					);
+	const handleSwatchStatus = useCallback(
+		async (e, idx) => {
+			await updateData.mutateAsync({
+				url: `/zipper/sfg-swatch-bulk-update/${data[idx]?.order_description_uuid}/${data[idx]?.style}/${data[idx]?.color}/${data[idx]?.bleaching}`,
+				updatedData: {
+					recipe_uuid: e.value,
+					swatch_approval_date: GetDateTime(),
 				},
-			},
-			{
-				accessorKey: 'item_description',
-				header: 'Item Description',
-				// enableColumnFilter: false,
-				width: 'w-32',
-				cell: (info) => {
-					const { order_description_uuid, order_number } =
-						info.row.original;
-					return (
-						<LinkWithCopy
-							title={info.getValue()}
-							id={order_description_uuid}
-							uri={`/order/details/${order_number}`}
-						/>
-					);
-				},
-			},
-			// {
-			// 	accessorKey: 'order_type',
-			// 	header: 'Type',
-			// 	width: 'w-24',
-			// 	// enableColumnFilter: false,
-			// 	cell: (info) => info.getValue(),
-			// },
-			{
-				accessorKey: 'style',
-				header: 'Style',
-				width: 'w-40',
-				// enableColumnFilter: false,
-				cell: (info) => info.getValue(),
-			},
-			{
-				accessorKey: 'color',
-				header: 'Color',
-				width: 'w-40',
-				// enableColumnFilter: false,
-				cell: (info) => info.getValue(),
-			},
-			{
-				accessorKey: 'color_ref',
-				header: 'Color Ref',
-				// enableColumnFilter: true,
-				cell: (info) => info.getValue(),
-			},
-			{
-				accessorKey: 'size',
-				header: 'Size',
-				width: 'w-24',
-				// enableColumnFilter: false,
-				cell: (info) => info.getValue(),
-			},
-			{
-				accessorFn: (row) => {
-					if (row.order_type === 'tape') return 'MTR';
-
-					return row.is_inch === 1 ? 'INCH' : 'CM';
-				},
-				id: 'unit',
-				header: 'Unit',
-				enableColumnFilter: false,
-			},
-			{
-				accessorKey: 'quantity',
-				header: (
-					<>
-						QTY <br />
-						(PCS)
-					</>
-				),
-				enableColumnFilter: false,
-				cell: (info) => info.getValue(),
-			},
-			{
-				accessorKey: 'bleaching',
-				header: 'Bleach',
-				enableColumnFilter: true,
-				enableSorting: true,
-				// cell: (info) => {
-				// 	const isBleach = info.getValue() === 'bleach';
-				// 	return (
-				// 		<StatusButton className={'btn-xs'} value={isBleach} />
-				// 	);
-				// },
-			},
-			{
-				accessorFn: (row) => (row.is_batch_created ? 'Yes' : 'No'),
-				id: 'is_dyeing_batch_entry',
-				header: (
-					<>
-						Batch <br />
-						Created
-					</>
-				),
-				enableColumnFilter: false,
-				cell: (info) => (
-					<StatusButton
-						className={'btn-xs'}
-						value={info.row.original.is_batch_created}
-					/>
-				),
-			},
-			{
-				accessorKey: 'receive_by_factory_time',
-				header: (
-					<>
-						Factory <br />
-						Received
-					</>
-				),
-				width: 'w-24',
-				enableColumnFilter: false,
-				cell: (info) => (
-					<DateTime
-						date={info.row.original.receive_by_factory_time}
-					/>
-				),
-			},
-			{
-				accessorFn: (row) =>
-					row.swatch_approval_received ? 'Yes' : 'No',
-				id: 'swatch_approval_received',
-				header: (
-					<>
-						Swatch <br />
-						Rcv.
-					</>
-				),
-				width: 'w-24',
-				enableColumnFilter: false,
-				cell: (info) => (
-					<div className='flex flex-col'>
-						<SwitchToggle
-							disabled={
-								!haveAccess.includes('click_swatch_status')
-							}
-							onChange={() => {
-								handleSwatchApprovalDate(info.row.index);
-							}}
-							checked={
-								info.row.original.swatch_approval_received ===
-								true
-							}
-						/>
-						{/* <DateTime date={info.row.original.bulk_approval_date} /> */}
-						<span>
-							{info.row.original.swatch_approval_received_by_name}
-						</span>
-					</div>
-				),
-			},
-			{
-				accessorFn: (row) => {
-					if (row.swatch_approval_received_date === null) return null;
-
-					return format(
-						row.swatch_approval_received_date,
-						'dd/MM/yyyy'
-					);
-				},
-				id: 'swatch_approval_received_date',
-				header: (
-					<>
-						Swatch App <br />
-						Rcv Date
-					</>
-				),
-				width: 'w-24',
-				enableColumnFilter: false,
-				cell: (info) => (
-					<DateTime
-						date={info.row.original.swatch_approval_received_date}
-					/>
-				),
-			},
-			{
-				accessorKey: 'recipe_name',
-				header: 'Recipe',
-				enableColumnFilter: false,
-				hidden: !haveAccess.includes('update'),
-				width: 'min-w-52',
-				cell: (info) => {
-					const { recipe_uuid } = info.row.original;
-					const { order_info_uuid, bleaching } = info.row.original;
-
-					const swatchAccess = haveAccess.includes(
-						'click_swatch_status'
-					);
-					const swatchAccessOverride = haveAccess.includes(
-						'click_swatch_status_override'
-					);
-
-					let isDisabled = true;
-					if (swatchAccessOverride) isDisabled = false;
-					if (recipe_uuid === null && swatchAccess)
-						isDisabled = false;
-
-					return (
-						<ReactSelect
-							key={recipe_uuid}
-							placeholder='Select order info uuid'
-							options={recipe?.filter(
-								(recipeItem) =>
-									(recipeItem.bleaching === bleaching &&
-										recipeItem.order_info_uuid ===
-											order_info_uuid) ||
-									recipeItem.value === null
-							)}
-							value={recipe?.find(
-								(item) => item.value == recipe_uuid
-							)}
-							onChange={(e) =>
-								handleSwatchStatus(e, info.row.index)
-							}
-							isDisabled={isDisabled}
-							menuPortalTarget={document.body}
-						/>
-					);
-				},
-			},
-			{
-				accessorFn: (row) => {
-					if (row.swatch_approval_date === null) return null;
-
-					return format(row.swatch_approval_date, 'dd/MM/yyyy');
-				},
-				id: 'swatch_approval_date',
-				header: (
-					<>
-						Setup <br />
-						Date
-					</>
-				),
-				width: 'w-24',
-				enableColumnFilter: false,
-				cell: (info) => (
-					<DateTime date={info.row.original.swatch_approval_date} />
-				),
-			},
-			{
-				accessorKey: 'color_ref_entry_date',
-				header: (
-					<>
-						Color Ref <br /> Entry
-					</>
-				),
-				filterFn: 'isWithinRange',
-				enableColumnFilter: false,
-				width: 'w-24',
-				cell: (info) => {
-					return <DateTime date={info.getValue()} />;
-				},
-			},
-			{
-				accessorKey: 'color_ref_update_date',
-				header: (
-					<>
-						Color Ref <br /> Update
-					</>
-				),
-				filterFn: 'isWithinRange',
-				enableColumnFilter: false,
-				width: 'w-24',
-				cell: (info) => {
-					return <DateTime date={info.getValue()} />;
-				},
-			},
-		],
-		[data, recipe, handleSwatchStatus, haveAccess]
+				isOnCloseNeeded: false,
+			});
+		},
+		[updateData, data]
 	);
 
-	if (isLoading)
+	const handleSwatchApprovalDate = useCallback(
+		async (idx) => {
+			await updateData.mutateAsync({
+				url: `/zipper/swatch-approval-received-bulk-update/${data[idx]?.order_description_uuid}/${data[idx]?.style}/${data[idx]?.color}/${data[idx]?.bleaching}`,
+				updatedData: {
+					swatch_approval_received: data[idx]
+						?.swatch_approval_received
+						? false
+						: true,
+					swatch_approval_received_date: data[idx]
+						?.swatch_approval_received
+						? null
+						: GetDateTime(),
+					swatch_approval_received_by: user?.uuid,
+				},
+				isOnCloseNeeded: false,
+			});
+		},
+		[updateData, data, user]
+	);
+
+	const handleSwatchApprovalDateLog = useCallback(
+		async (idx) => {
+			await updateData.mutateAsync({
+				url: `/zipper/swatch-approval-received/${data[idx]?.uuid}`,
+				updatedData: {
+					swatch_approval_received: data[idx]
+						?.swatch_approval_received
+						? false
+						: true,
+					swatch_approval_received_date: data[idx]
+						?.swatch_approval_received
+						? null
+						: GetDateTime(),
+					swatch_approval_received_by: user?.uuid,
+				},
+				isOnCloseNeeded: false,
+			});
+		},
+		[updateData, data, user]
+	);
+
+	// Create columns using the factory functions
+	const columns = useMemo(
+		() =>
+			createColumns({
+				data: bulkSwatch,
+				recipe,
+				handleSwatchStatus,
+				haveAccess,
+				handleSwatchApprovalDate,
+				user,
+			}),
+		[
+			bulkSwatch,
+			recipe,
+			handleSwatchStatus,
+			haveAccess,
+			handleSwatchApprovalDate,
+			user,
+		]
+	);
+
+	const columnsLog = useMemo(
+		() =>
+			createColumnsLog({
+				data,
+				recipe,
+				handleSwatchStatusLog,
+				haveAccess,
+				handleSwatchApprovalDateLog,
+				user,
+			}),
+		[
+			data,
+			recipe,
+			handleSwatchStatusLog,
+			haveAccess,
+			handleSwatchApprovalDateLog,
+			user,
+		]
+	);
+
+	if (isLoading && isLoadingBulkSwatch)
 		return <span className='loading loading-dots loading-lg z-50' />;
 
 	return (
-		<div>
+		<div className='flex flex-col gap-2'>
 			<ReactTable
 				title={info.getTitle()}
-				data={data}
+				data={bulkSwatch}
 				columns={columns}
+				extraButton={
+					<>
+						{status2 === 'incomplete_order' && (
+							<StatusSelect
+								options={options}
+								status={status}
+								setStatus={setStatus}
+							/>
+						)}
+						<StatusSelect
+							options={options2}
+							status={status2}
+							setStatus={setStatus2}
+						/>
+					</>
+				}
+			/>
+			<hr className='my-6 border-2 border-dashed border-secondary-content bg-black' />
+			<ReactTable
+				title={'Log'}
+				data={data}
+				columns={columnsLog}
 				extraButton={
 					<>
 						{status2 === 'incomplete_order' && (
